OvmfPkg/TdxHelperLib: Read and Validate TDX Metadata
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=4363

The TDX Metadata is the descriptor for TDVF, it provides information to
the VMM that is used to build a TD. The VMM need to use an offset to find
the metadata. This offset is on the GUID Struct Chain that is defined on
the reset vector (OvmfPkg/ResetVector/Ia16/ResetVectorVtf0.asm).
Both the chain and metadata are stored in the BFV.

In this patch, tdvf first reads the offset from the chain, then uses the
offset to read and validate the tdx metadata.

Cc: Erdem Aktas <erdemaktas@google.com>
Cc: James Bottomley <jejb@linux.ibm.com>
Cc: Jiewen Yao <jiewen.yao@intel.com>
Cc: Min Xu <min.m.xu@intel.com>
Cc: Tom Lendacky <thomas.lendacky@amd.com>
Cc: Michael Roth <michael.roth@amd.com>
Signed-off-by: Ceping Sun <cepingx.sun@intel.com>