OvmfPkg/TdxHelperLib: Check Unaccepted Memory region of TD HOB
REF: https://bugzilla.tianocore.org/show_bug.cgi?id=4363

VMM is not in the TCB, so VMM may manipulate the TD HOB with incorrect
resource descriptor hob of memory regions. For example, it may replace the
unaccepted memory region with private memory region so that TDVF would
accept the private pages again. That is considered as an attack.

Currently,the unaccepted memory region of TD HOB is not checked and the
initial private page is recorded on the TDX metadata.

So, before TDVF accepts the unaccepted memory described in TD HOB, the
unaccepted memory regions are checked based on the temporary memory region
of TDX metadata (TDX_METADATA_SECTION_TYPE_TEMP_MEM).

Cc: Erdem Aktas <erdemaktas@google.com>
Cc: James Bottomley <jejb@linux.ibm.com>
Cc: Jiewen Yao <jiewen.yao@intel.com>
Cc: Min Xu <min.m.xu@intel.com>
Cc: Tom Lendacky <thomas.lendacky@amd.com>
Cc: Michael Roth <michael.roth@amd.com>
Signed-off-by: Ceping Sun <cepingx.sun@intel.com>